Thanks @bohendo for a lovely entry.sh script
Removed certbot container. Too complicated.

The reason is because we need a way of generating dummy
certificates and it is easier to just do it all inside of
the nginx container than trying to make it in a seperate
certbot container

Fix typo

Fix another typo

Remove setup rule

Remove leftover setup rule

Again remove setup